requirements.txt
fastapi~=0.108.0
pydantic~=1.10.11
starlette==0.30.0
numpy~=1.26.3
tensorflow~=2.15.0
keras~=2.15.0
pillow~=10.2.0
python-dotenv~=1.0.0